requirements.txt
zenml
pandas
easyocr
spacy
mlflow
python-multipart
Pillow
PyPDF2
pdf2image
pydantic
typer
rich
flask
python-dotenv
google-generativeai
markdown
